Updated data types for MPL115A2 temp and press

+++ src/gen_dragonfruit/gd_dev_adafruit_MPL115A2_press.h
@@ -3,5 +3,5 @@
 #define _PIN_GD_ADAFRUIT_MPL115A2_H A4
 #ifndef _PIN_GD_ADAFRUIT_MPL115A2_H
 void gd_dev_adafruit_MPL115A2_press_open(void);
-int gd_dev_adafruit_MPL115A2_press_read(void);
+uint16_t gd_dev_adafruit_MPL115A2_press_read(void);
 #endif
